backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy==2.0.36
pyautogen==0.2.23
google-generativeai==0.8.1
chromadb==0.5.20
sentence-transformers==3.3.1
duckduckgo-search==8.2.1
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.6.1
aiofiles==24.1.0
python-dotenv==1.0.1
alembic==1.15.0
beautifulsoup4==4.12.3
requests==2.32.3
openai==1.58.1

